uint8_t replaced

+++ firmware/Metro-Arduino-Spark-Wiring.cpp
@@ -54,7 +54,7 @@ void Metro::interval(unsigned long interval_millis)
   this->interval_millis = interval_millis;
 }
 
-uint8_t Metro::check()
+unsigned int Metro::check()
 {
 
   unsigned long now = millis();
